Renamed packages from java.main org.mftech
