+++ src/org/mftech/Launcher.java
@@ -1,4 +1,4 @@
-package main.java;
+package org.mftech;
 
 public class Launcher {
 	public static void main(String[] args) {
