+++ src/org/mftech/Model/Enumerations/ActionType.java
@@ -1,4 +1,4 @@
-package main.java.Model.Enumerations;
+package org.mftech.Model.Enumerations;
 
 public enum ActionType {
 	Raise,
