+++ src/org/mftech/Model/Enumerations/CombinationType.java
@@ -1,4 +1,4 @@
-package main.java.Model.Enumerations;
+package org.mftech.Model.Enumerations;
 
 public enum CombinationType {
 	NoCombination (0),
