+++ src/org/mftech/Model/Enumerations/StageType.java
@@ -1,4 +1,4 @@
-package main.java.Model.Enumerations;
+package org.mftech.Model.Enumerations;
 
 public enum StageType {
 	PreFlop		(0),
