+++ src/org/mftech/Model/HistoryBet.java
@@ -1,7 +1,7 @@
-package main.java.Model;
+package org.mftech.Model;
 
-import main.java.Model.Enumerations.ActionType;
-import main.java.Model.Enumerations.StageType;
+import org.mftech.Model.Enumerations.ActionType;
+import org.mftech.Model.Enumerations.StageType;
 
 public class HistoryBet {
 	public final StageType type;
