+++ src/org/mftech/Utility/ConsoleHelper.java
@@ -1,4 +1,4 @@
-package main.java.Utility;
+package org.mftech.Utility;
 import java.io.BufferedReader;
 
 import java.io.IOException;
