+++ src/tests/java/CombinationTest.java
@@ -1,19 +1,17 @@
 package tests.java;
 
-import main.java.Model.AiPlayer;
-import main.java.Model.Card;
-import main.java.Model.Enumerations.CardType;
-import main.java.Model.Enumerations.CardValue;
-import main.java.Model.Enumerations.CombinationType;
-import main.java.Model.Enumerations.RaisingPolicy;
-import main.java.Model.Player;
-import main.java.Model.Table;
-import main.java.Model.TableSettings;
-
 import org.junit.Before;
 import org.junit.Test;
-
-import AiStrategies.BasicStrategy;
+import org.mftech.AiStrategies.BasicStrategy;
+import org.mftech.Model.AiPlayer;
+import org.mftech.Model.Card;
+import org.mftech.Model.Player;
+import org.mftech.Model.Table;
+import org.mftech.Model.TableSettings;
+import org.mftech.Model.Enumerations.CardType;
+import org.mftech.Model.Enumerations.CardValue;
+import org.mftech.Model.Enumerations.CombinationType;
+import org.mftech.Model.Enumerations.RaisingPolicy;
 
 import static org.junit.Assert.*;
